# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "EMS API Documentation",
        "description": "Dokumentasi API untuk event management system:<br>Auth: Register, Login, dan Logout<br>Event: daftar event, detail event, pendaftaran user<br>Attendance: check-in, check-out, dan riwayat",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Server utama API EMS"
        }
    ],
    "paths": {
...
        "/api/v1/user/profile": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Ambil profil user yang sedang login",
                "description": "Mengembalikan data profil user yang sedang login berdasarkan token autentikasi.",
                "operationId": "cc344fec551903e2b2cbc6f5d5fa7b5f",
                "responses": {
                    "200": {
                        "description": "Data profil user",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Damar"
                                        },
                                        "email": {
                                            "type": "string",
                                            "example": "damar@example.com"
                                        },
                                        "phone": {
                                            "type": "string",
                                            "example": "+628123456789"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "datetime",
                                            "example": "2025-10-16T09:00:00Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/user/update-profile": {
            "put": {
                "tags": [
                    "User"
                ],
                "summary": "Update profil user",
                "description": "Memperbarui data profil user seperti nama dan nomor telepon.",
                "operationId": "f53ccb591303131db6695422ff4027e1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "maxLength": 100,
                                        "example": "Damar Setiawan"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "maxLength": 15,
                                        "example": "+628123456789"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Profil berhasil diperbarui",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Damar Setiawan"
                                        },
                                        "email": {
                                            "type": "string",
                                            "example": "damar@example.com"
                                        },
                                        "phone": {
                                            "type": "string",
                                            "example": "+628123456789"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/user/change-password": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Ubah password user",
                "description": "Mengganti password user dengan validasi password lama.",
                "operationId": "51a9fe42c70c2d3e5689358712d06f93",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "old_password",
                                    "new_password"
                                ],
                                "properties": {
                                    "old_password": {
                                        "type": "string",
                                        "example": "oldpass123"
                                    },
                                    "new_password": {
                                        "type": "string",
                                        "minLength": 6,
                                        "example": "newsecurepassword"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Password berhasil diperbarui",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Password updated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Password lama salah",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Incorrect old password"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}